requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
arrow-cpp=8.0.0=py39h60b952e_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py39hb9d737c_1004
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.0=py39h4bc2ebd_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cryptography=3.4.8=py39h95dcef6_1
dataclasses=0.8=pyhc8e2a94_3
datasets=2.7.1=pyhd8ed1ab_0
dill=0.3.6=pyhd8ed1ab_1
filelock=3.9.0=pyhd8ed1ab_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2022.11.0=pyhd8ed1ab_0
future=0.18.2=pyhd8ed1ab_6
gflags=2.2.2=he1b5a44_1004
glog=0.6.0=h6f12383_0
grpc-cpp=1.46.1=h33aed49_0
huggingface_hub=0.11.1=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_metadata=6.0.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcurl=7.86.0=h91b91d3_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.46.0=hce63b2e_0
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=he6d91bd_0
lz4-c=1.9.3=h9c3ff4c_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
multidict=6.0.2=py39h5eee18b_0
multiprocess=0.70.12.2=py39hb9d737c_2
ncurses=6.3=h5eee18b_3
ninja=1.11.0=h924138e_0
numexpr=2.8.4=py39he184ba9_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
openssl=1.1.1s=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=22.0=pyhd8ed1ab_0
pandas=1.5.2=py39h417a72b_0
pip=22.3.1=py39h06a4308_0
pyarrow=8.0.0=py39h992f0b0_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=3.0.0=py39hb9d737c_1
python_abi=3.9=2_cp39
pytorch=1.12.1=cpu_py39he8d8e81_0
pytz=2022.7=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_4
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py39h5eee18b_0
requests=2.28.1=pyhd8ed1ab_1
responses=0.18.0=pyhd8ed1ab_0
sacremoses=0.0.53=pyhd8ed1ab_0
setuptools=65.5.0=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=hbd366e4_1
sqlite=3.40.0=h5082296_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py39h3dcd8bd_1
tqdm=4.64.1=pyhd8ed1ab_0
transformers=4.24.0=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h04d1e81_0
unidecode=1.3.6=pyhd8ed1ab_0
urllib3=1.26.13=pyhd8ed1ab_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=0.8.0=h7f98852_3
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py39hb9d737c_2
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
